requirements.txt
SpeechRecognition==3.10.0
pyttsx3==2.90
python-dotenv==1.0.0
requests==2.31.0
azure-cognitiveservices-speech==1.34.0
openai==1.4.0
fpdf==1.7.2
argparse==1.4.0 